fix: state fixes
