+++ src/store/common/api-clients.ts
@@ -9,7 +9,7 @@ import {
 } from '@stacks/blockchain-api-client';
 import { fetcher } from '@common/api/wrapped-fetch';
 
-enum KEYS {
+enum API_CLIENT_KEYS {
   CONFIG = 'clients/CONFIG',
   SMART_CONTRACTS = 'clients/SMART_CONTRACTS',
   ACCOUNTS = 'clients/ACCOUNTS',
@@ -18,7 +18,7 @@ enum KEYS {
 }
 
 export const apiClientConfiguration = selector({
-  key: KEYS.CONFIG,
+  key: API_CLIENT_KEYS.CONFIG,
   get: ({ get }) => {
     const network = get(currentNetworkState);
     return new Configuration({ basePath: network.url, fetchApi: fetcher });
@@ -26,7 +26,7 @@ export const apiClientConfiguration = selector({
 });
 
 export const smartContractClientState = selector({
-  key: KEYS.SMART_CONTRACTS,
+  key: API_CLIENT_KEYS.SMART_CONTRACTS,
   get: ({ get }) => {
     const config = get(apiClientConfiguration);
     return new SmartContractsApi(config);
@@ -34,7 +34,7 @@ export const smartContractClientState = selector({
 });
 
 export const accountsApiClientState = selector({
-  key: KEYS.ACCOUNTS,
+  key: API_CLIENT_KEYS.ACCOUNTS,
   get: ({ get }) => {
     const config = get(apiClientConfiguration);
     return new AccountsApi(config);
@@ -42,7 +42,7 @@ export const accountsApiClientState = selector({
 });
 
 export const infoApiClientState = selector({
-  key: KEYS.INFO,
+  key: API_CLIENT_KEYS.INFO,
   get: ({ get }) => {
     const config = get(apiClientConfiguration);
     return new InfoApi(config);
@@ -50,7 +50,7 @@ export const infoApiClientState = selector({
 });
 
 export const blocksApiClientState = selector({
-  key: KEYS.BLOCKS,
+  key: API_CLIENT_KEYS.BLOCKS,
   get: ({ get }) => {
     const config = get(apiClientConfiguration);
     return new BlocksApi(config);
